requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pymupdf
requests
jinja2
python-multipart
streamlit